# docs/design/openapi.yaml
openapi: 3.1.0
info:
  title: KACP API
  version: 0.1.0-draft
  description: |
    KCC Agent Cloud Platform 외부 API (0단계 초안).
    규칙과 내부 API는 docs/design/04-api.md 참고.
    상태 변경 요청은 X-KACP-CSRF 헤더가 필요하다.
servers:
  - url: https://app.kacp.cloud/api/v1
security:
  - session: []
...
paths:
...
  /departments:
    get:
      tags: [org]
      summary: 부서 트리 (선택 화면용)
      parameters:
        - { name: includeArchived, in: query, schema: { type: boolean, default: false }, description: 관리자만 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/DepartmentNode' }
...
  /admin/departments:
    post:
      tags: [admin]
      parameters: [{ $ref: '#/components/parameters/Csrf' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DepartmentInput' }
      responses:
        '201':
          description: 생성됨
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Department' }
        '409': { $ref: '#/components/responses/Error' }
  /admin/departments/{id}:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    patch:
      tags: [admin]
      parameters: [{ $ref: '#/components/parameters/Csrf' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DepartmentInput' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Department' }
  /admin/departments/{id}/move:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    post:
      tags: [admin]
      parameters: [{ $ref: '#/components/parameters/Csrf' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [parentId]
              properties:
                parentId: { type: [string, 'null'], format: uuid }
                sortOrder: { type: integer }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Department' }
        '422': { $ref: '#/components/responses/Error' }
  /admin/departments/{id}/{action}:
    parameters:
      - { $ref: '#/components/parameters/Id' }
      - { name: action, in: path, required: true, schema: { type: string, enum: [archive, unarchive] } }
    post:
      tags: [admin]
      parameters: [{ $ref: '#/components/parameters/Csrf' }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Department' }
        '409': { $ref: '#/components/responses/Error' }
  /admin/departments/{id}/members:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    get:
      tags: [admin]
      parameters:
        - { name: includeDescendants, in: query, schema: { type: boolean, default: false } }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/AdminUser' }
  /admin/import/{kind}/preview:
    parameters:
      - { name: kind, in: path, required: true, schema: { type: string, enum: [departments, users] } }
    post:
      tags: [admin]
      summary: CSV 미리보기 (저장하지 않음)
      parameters: [{ $ref: '#/components/parameters/Csrf' }]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: { type: string, format: binary }
      responses:
        '200':
          description: 미리보기
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ImportJob' }
        '422': { $ref: '#/components/responses/Error' }
  /admin/import/{kind}/apply:
    parameters:
      - { name: kind, in: path, required: true, schema: { type: string, enum: [departments, users] } }
    post:
      tags: [admin]
      parameters: [{ $ref: '#/components/parameters/Csrf' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [jobId]
              properties:
                jobId: { type: string, format: uuid }
                skipErrors: { type: boolean, default: false }
      responses:
        '200':
          description: 적용 결과
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ImportJob'
                  - type: object
                    properties:
                      initialPasswordsCsv:
                        type: [string, 'null']
                        description: users 가져오기에서 새로 만든 계정의 초기 비밀번호 CSV. 이 응답에서만 한 번 제공
        '409': { $ref: '#/components/responses/Error' }
...
components:
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: kacp_session
  parameters:
    Csrf:
      name: X-KACP-CSRF
      in: header
      required: true
      schema: { type: string }
...
    Id:
      name: id
      in: path
      required: true
      schema: { type: string }
...
  responses:
    Error:
      description: 오류
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string, examples: [NAME_TAKEN] }
            message: { type: string, description: 사용자에게 보여줄 한국어 문구 }
            details: { type: object }
...
    PlatformRole:
      type: string
      enum: [admin, user]
    TeamRole:
      type: string
      enum: [team_admin, member]
...
    UserRef:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        email: { type: string }
        departmentName: { type: [string, 'null'], description: 동명이인 구분용 표시 }
    DepartmentRef:
      type: object
      properties:
        id: { type: string, format: uuid }
        code: { type: string }
        name: { type: string }
        pathNames: { type: array, items: { type: string }, description: "최상위부터 (예: [경영지원본부, 인사팀])" }
    Department:
      allOf:
        - $ref: '#/components/schemas/DepartmentRef'
        - type: object
          properties:
            parentId: { type: [string, 'null'], format: uuid }
            depth: { type: integer }
            sortOrder: { type: integer }
            head: { oneOf: [{ $ref: '#/components/schemas/UserRef' }, { type: 'null' }] }
            status: { type: string, enum: [active, archived] }
            memberCount: { type: integer, description: 이 부서만 }
            totalMemberCount: { type: integer, description: 하위 포함 }
    DepartmentNode:
      allOf:
        - $ref: '#/components/schemas/Department'
        - type: object
          properties:
            children:
              type: array
              items: { $ref: '#/components/schemas/DepartmentNode' }
    DepartmentInput:
      type: object
      properties:
        name: { type: string }
        code: { type: string }
        parentId: { type: [string, 'null'], format: uuid }
        headUserId: { type: [string, 'null'], format: uuid }
        sortOrder: { type: integer }
    ImportJob:
      type: object
      properties:
        id: { type: string, format: uuid }
        kind: { type: string, enum: [departments, users] }
        status: { type: string, enum: [previewed, applied, cancelled] }
        fileName: { type: string }
        summary:
          type: object
          properties:
            added: { type: integer }
            updated: { type: integer }
            unchanged: { type: integer }
            errors: { type: integer }
        rows:
          type: array
          items:
            type: object
            properties:
              line: { type: integer }
              action: { type: string, enum: [add, update, unchanged, error] }
              key: { type: string, description: code 또는 email }
              changes: { type: object, description: "바뀌는 칸 {field: [before, after]}" }
              error: { type: [string, 'null'] }
...
    AdminUser:
      type: object
      properties:
        id: { type: string, format: uuid }
        email: { type: string }
        name: { type: string }
        department: { oneOf: [{ $ref: '#/components/schemas/DepartmentRef' }, { type: 'null' }] }
        title: { type: [string, 'null'] }
        employeeNo: { type: [string, 'null'] }
        platformRole: { $ref: '#/components/schemas/PlatformRole' }
        status: { type: string, enum: [active, disabled] }
        mustChangePassword: { type: boolean }
        teams:
          type: array
          items:
            type: object
            properties:
              name: { type: string }
              teamRole: { $ref: '#/components/schemas/TeamRole' }
        lastLoginAt: { type: [string, 'null'], format: date-time }
        createdAt: { type: string, format: date-time }
...